A setting derived in Python gets rounded the way a board would (audioif#80)
clean-build.yml's unix-usermod (float-precision) cell builds the unix port
with -DMICROPY_FLOAT_IMPL=MICROPY_FLOAT_IMPL_FLOAT, which is what every
board we ship is, and verify_dsp had six probes disagreeing with CPython
there. Reproduced locally at v1.29.0 against this worktree, same six, same
byte offsets. Four of the six were the probes' own arithmetic.

Python's float is the interpreter's mp_float_t. `node.mix = 0.35` is
0.34999999403953552 on a board and 0.34999999999999998 here, so the number
was already two numbers before any kernel saw it. lib/audioif_util is
struct and two functions: float32(value), the round trip that is the
identity on a single-precision target and a rounding on a double one, and
float32_bits(value), the exact way to print a float two interpreters have
to agree on. docs/correctness-standard.md carries the rule -- a setting
derived in Python passes through float32 before it reaches a node -- and
says plainly that it is what audiocomponents#75 needs on the class side.

Per probe, with what proves it:

dynamics_probe, dynamics_extras_probe, dynamics_options_probe: 8, 3 and 24
lines, and on EVERY one of them only the trailing "%.6f" gain reduction
differed -- byte sum and FNV checksum identical, so the audio never moved.
The value does not move either: the float32 bits are f2c7bac0 on CPython,
on the float build and on the double build alike. Six decimals of a single
is seven significant digits and MicroPython's single-precision formatter is
not correctly rounded that far, so that column was comparing formatters.
They print float32_bits now, which is exact and strictly more sensitive.

granular_pitch_shift_probe: one case, mix=0.35. float32(0.35) and the probe
agrees, 36 lines of 36. spread=0.6 goes through it too though it happened
not to move.

echo_filter_probe: decay=0.7 narrows the same way and was the whole of the
un-filtered case. Fixed; what is left is audioif#102.

biquad_component_probe: 252 of 286 lines, and the cause was the fixture's
own material. `(2*pi*hz*frame/RATE) % 2*pi < pi` in Python floats is a
DIFFERENT SQUARE WAVE on a single-precision build -- 96000 against 120000
summed in mono, 8 frames flipped there and 96 in stereo, frame 400 landing
at exactly 0.0 on a double and wrapping to 6.2831052 on a single. So this
gate was comparing two kernels on two signals. The material is integers
now; Q and A go through float32; 248 of the 252 lines close and what is
left is audioif#101. The integer form is not byte-identical to the float
one on a double build either (the float form classified a handful of exact
crossings the other way), so golden/biquad_component.json is re-captured:
7d1bbf0d... -> d9950b16..., an intended fixture change, which is what that
file's --capture exists for.

The two that are left are below the probes and each got its own issue,
because each moves board digests and wants a firmware rebuild and a board
proof rather than a quiet landing:

  audioif#101  src/synthio/Biquad.c:79 derives W0 = frequency *
               synthio_global_W_scale in mp_float_t, where the CPython
               extension calls the shared double audioif_biquad_cp_w0()
               (_audioif.c:588). At 1200 Hz / 8 kHz that is
               0.942477796076938 against 0.9424778819084167, and Q15
               coefficient rounding turns it into a different filter.
               Proof: feeding the twin frequency 1200.000109284033, whose
               double derivation lands on exactly the float build's W0,
               reproduces that build's bytes -- 63659 3365258812 and
               63704 1319762287, both blocks.

  audioif#102  A filtered audiodelays.Echo computes echo * decay +
               sample_word in mp_float_t (Echo.c:389, :395) and in Python
               double in the twin (audiodelays.py:82, :88). Proof:
               rounding those two expressions to float32 in the twin makes
               the probe identical, 34 lines of 34.

The CI cell's --known-divergent list drops from six probes to those two,
with the issue numbers beside it, so a third probe diverging still fails
and either of these agreeing still fails.

Gates. verify_dsp on the float32 build: 25 comparisons, 2 failures, both
XFAIL, "Gate satisfied". On cmods/bin/micropython: 25 comparisons, 0
failures. Three-way with circuitpython-oracle-10.3.0: every probe before
samplehold_probe agrees three ways, and samplehold_probe aborts the run on
that binary because the pinned oracle predates audioshaper.SampleHold --
which main does too, checked, so it is not this change. The three probes
changed here that run after it were three-wayed by hand and agree.
unittest discover: 313 tests, OK (1 skipped). The five in-repo parity
gates pass. flake8 clean. The wheel carries audioif_util/__init__.py.

+++ tests/parity/granular_pitch_shift_probe.py
@@ -10,6 +10,7 @@ import sys
 from array import array
 
 import audiocore
+from audioif_util import float32
 
 MODULE = sys.argv[1] if len(sys.argv) > 1 else "audiodelays"
 delays = __import__(MODULE)
@@ -47,10 +48,17 @@ CASES = (
             "spread": 0.0}),
     ("down", {"semitones": -7, "mix": 1.0, "grain_size": 48, "density": 3,
               "spread": 0.0}),
-    ("blend", {"semitones": 3, "mix": 0.35, "grain_size": 64, "density": 2,
-               "spread": 0.0}),
+    # `float32(0.35)` rather than 0.35: the node keeps `mix` as an
+    # `mp_float_t`, so a plain literal is one number on a double build and
+    # another on a single-precision one, and the blend renders different
+    # bytes before the kernel is reached at all. That was one of the six
+    # divergences in audioif#80; `spread` below goes through it for the same
+    # reason even though it happened not to move. 0.0, 0.5 and 1.0 are exact
+    # in both widths and need no such call.
+    ("blend", {"semitones": 3, "mix": float32(0.35), "grain_size": 64,
+               "density": 2, "spread": 0.0}),
     ("cloud", {"semitones": 2, "mix": 1.0, "grain_size": 80, "density": 4,
-               "spread": 0.6}),
+               "spread": float32(0.6)}),
 )
 
 for name, options in CASES:
